+++ src/GitHub_Updater/Traits/Basic_Auth_Loader.php
@@ -14,6 +14,9 @@ use Fragen\Singleton;
 use Fragen\GitHub_Updater\Install;
 use Fragen\GitHub_Updater\API\Bitbucket_API;
 use Fragen\GitHub_Updater\API\Bitbucket_Server_API;
+use Fragen\GitHub_Updater\API\GitHub_API;
+use Fragen\GitHub_Updater\API\GitLab_API;
+use Fragen\GitHub_Updater\API\Gitea_API;
 
 /*
  * Exit if called directly.
@@ -31,7 +34,7 @@ trait Basic_Auth_Loader {
 	 *
 	 * @var array
 	 */
-	private static $basic_auth_required = [ 'Bitbucket' ];
+	private static $basic_auth_required = [ 'Bitbucket', 'GitHub', 'GitLab', 'Gitea' ];
 
 	/**
 	 * Load hooks for Bitbucket authentication headers.
@@ -54,8 +57,8 @@ trait Basic_Auth_Loader {
 	}
 
 	/**
-	 * Add Basic Authentication $args to http_request_args filter hook
-	 * for private repositories only.
+	 * Add Basic Authentication $args to http_request_args filter hook.
+	 * Bitbucket private repositories only.
 	 *
 	 * @access public
 	 *
@@ -66,14 +69,23 @@ trait Basic_Auth_Loader {
 	 */
 	public function maybe_basic_authenticate_http( $args, $url ) {
 		$credentials = $this->get_credentials( $url );
-
-		if ( $credentials['private'] && $credentials['isset'] && ! $credentials['api.wordpress'] ) {
+		if ( ! $credentials['isset'] || $credentials['api.wordpress'] ) {
+			return $args;
+		}
+		if ( 'bitbucket' === $credentials['type'] ) {
 			$username = $credentials['username'];
 			$password = $credentials['password'];
 
 			// phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_encode
 			$args['headers']['Authorization'] = 'Basic ' . base64_encode( "$username:$password" );
 		}
+		if ( 'github' === $credentials['type'] || 'gitea' === $credentials['type'] ) {
+			$args['headers']['Authorization'] = 'token ' . $credentials['token'];
+		}
+		if ( 'gitlab' === $credentials['type'] ) {
+			$args['headers']['Authorization'] = 'Bearer ' . $credentials['token'];
+		}
+
 		remove_filter( 'http_request_args', [ $this, 'maybe_basic_authenticate_http' ] );
 
 		return $args;
@@ -89,6 +101,7 @@ trait Basic_Auth_Loader {
 	 * @return array $credentials
 	 */
 	private function get_credentials( $url ) {
+		$options      = get_site_option( 'github_updater' );
 		$headers      = parse_url( $url );
 		$username_key = null;
 		$password_key = null;
@@ -98,16 +111,21 @@ trait Basic_Auth_Loader {
 			'api.wordpress' => 'api.wordpress.org' === $headers['host'],
 			'isset'         => false,
 			'private'       => false,
+			'token'         => null,
+			'type'          => null,
 		];
 		$hosts        = [ 'bitbucket.org', 'api.bitbucket.org' ];
+
+		if ( $credentials['api.wordpress'] ) {
+			return $credentials;
+		}
 
 		$repos = array_merge(
 			Singleton::get_instance( 'Plugin', $this )->get_plugin_configs(),
 			Singleton::get_instance( 'Theme', $this )->get_theme_configs()
 		);
-
-		$slug = $this->get_slug_for_credentials( $headers, $repos, $url );
-		$type = $this->get_type_for_credentials( $slug, $repos, $url );
+		$slug  = $this->get_slug_for_credentials( $headers, $repos, $url, $options );
+		$type  = $this->get_type_for_credentials( $slug, $repos, $url );
 
 		switch ( $type ) {
 			case 'bitbucket':
@@ -116,18 +134,37 @@ trait Basic_Auth_Loader {
 				$bitbucket_org = in_array( $headers['host'], $hosts, true );
 				$username_key  = $bitbucket_org ? 'bitbucket_username' : 'bitbucket_server_username';
 				$password_key  = $bitbucket_org ? 'bitbucket_password' : 'bitbucket_server_password';
+				$type          = 'bitbucket';
 				break;
+			case 'github':
+			case $type instanceof GitHub_API:
+				$token = ! empty( $options[ $slug ] ) ? $options[ $slug ] : $options['github_access_token'];
+				$type  = 'github';
+				break;
+			case 'gitlab':
+			case $type instanceof GitLab_API:
+				$token = ! empty( $options[ $slug ] ) ? $options[ $slug ] : $options['gitlab_access_token'];
+				$type  = 'gitlab';
+				break;
+			case 'gitea':
+			case $type instanceof Gitea_API:
+				$token = ! empty( $options[ $slug ] ) ? $options[ $slug ] : $options['gitea_access_token'];
+				$type  = 'gitea';
 		}
 
-		// TODO: can use `( $this->caller )::$options` in PHP7.
-		$caller          = $this->get_class_vars( 'Base', 'caller' );
-		static::$options = $caller instanceof Install ? $caller::$options : static::$options;
-
-		if ( isset( static::$options[ $username_key ], static::$options[ $password_key ] ) ) {
-			$credentials['username'] = static::$options[ $username_key ];
-			$credentials['password'] = static::$options[ $password_key ];
+		if ( 'bitbucket' === $type && isset( $options[ $username_key ], $options[ $password_key ] ) ) {
+			$credentials['username'] = $options[ $username_key ];
+			$credentials['password'] = $options[ $password_key ];
 			$credentials['isset']    = true;
 			$credentials['private']  = $this->is_repo_private( $url );
+			$credentials['type']     = $type;
+		}
+
+		if ( 'github' === $type || 'gitlab' === $type || 'gitea' === $type ) {
+			$credentials['isset']   = true;
+			$credentials['private'] = $this->is_repo_private( $url );
+			$credentials['type']    = $type;
+			$credentials['token']   = isset( $token ) ? $token : null;
 		}
 
 		return $credentials;
@@ -139,12 +176,14 @@ trait Basic_Auth_Loader {
 	 * @param array  $headers Array of headers from parse_url().
 	 * @param array  $repos   Array of repositories.
 	 * @param string $url     URL being called by API.
+	 * @param array  $options Array of site options.
 	 *
 	 * @return bool|string $slug
 	 */
-	private function get_slug_for_credentials( $headers, $repos, $url ) {
+	private function get_slug_for_credentials( $headers, $repos, $url, $options ) {
 		$slug = isset( $_REQUEST['slug'] ) ? $_REQUEST['slug'] : false;
 		$slug = ! $slug && isset( $_REQUEST['plugin'] ) ? $_REQUEST['plugin'] : $slug;
+		$slug = false !== strpos( $slug, '/' ) ? dirname( $slug ) : $slug;
 		$slug = ! $slug && isset( $_REQUEST['theme'] ) ? $_REQUEST['theme'] : $slug;
 
 		// Set for bulk upgrade.
@@ -171,7 +210,8 @@ trait Basic_Auth_Loader {
 		if ( ! $slug && isset( $headers['path'] ) ) {
 			$path_arr = explode( '/', $headers['path'] );
 			foreach ( $path_arr as $key ) {
-				if ( array_key_exists( $key, $repos ) ) {
+				$key = basename( rawurldecode( $key ) ); // For GitLab.
+				if ( ! empty( $options[ $key ] ) || array_key_exists( $key, $repos ) ) {
 					$slug = $key;
 					break;
 				}
@@ -258,7 +298,7 @@ trait Basic_Auth_Loader {
 	}
 
 	/**
-	 * Removes Basic Authentication header for Bitbucket Release Assets.
+	 * Removes Basic Authentication header for Release Assets.
 	 * Storage in AmazonS3 buckets, uses Query String Request Authentication Alternative.
 	 *
 	 * @access public
@@ -270,8 +310,10 @@ trait Basic_Auth_Loader {
 	 * @return array $args
 	 */
 	public function http_release_asset_auth( $args, $url ) {
-		$arr_url = parse_url( $url );
-		if ( isset( $arr_url['host'] ) && 'bbuseruploads.s3.amazonaws.com' === $arr_url['host'] ) {
+		$arr_url         = parse_url( $url );
+		$aws_host        = false !== strpos( $arr_url['host'], 's3.amazonaws.com' );
+		$github_releases = false !== strpos( $arr_url['path'], 'releases/download' );
+		if ( $aws_host || $github_releases ) {
 			unset( $args['headers']['Authorization'] );
 		}
 		remove_filter( 'http_request_args', [ $this, 'http_release_asset_auth' ] );