+++ CHANGES.md
@@ -1,5 +1,12 @@
 #### [unreleased]
 
+#### 9.3.0 / 2020-02-06
+* remove GitHub deprecation notice
+* transition from GitHub access token query arg to Basic Authentication
+* fixed theme update View details display [#849](https://github.com/afragen/github-updater/issues/849)
+* more fixes PHP 7.4 warnings
+* refactor from using access token endpoints to Basic Authentication headers
+
 #### 9.2.4 / 2020-02-04
 * add notice re: GitHub deprecation notice 🤬